refactor(github): lift read actions into actions/reads.ts module
Pure refactor — the 9 read actions and their pagination helper move
out of index.ts into a buildReadActions() factory under
src/connectors/github/actions/reads.ts. index.ts now imports + spreads
the result. Establishes the per-domain action-module pattern that
PRs, issues, comments, releases, and workflows will follow.
